Improving the drag and drop
We now have (pretty ugly) drop areas for targeting the thingies.
They only get shown where they make sense. The is now a drag handler.
It only gets shown, when there is another position to drag to. We abuse
the existing arrow up and down handlers. If one of those is visible we
know we can change positions.

I think it would be nice to have some animation for the dropboxes to
appear. Also, we can change the cursor during drag. And the dropboxes
are ugly.